Add tests to components created and modified
Add text in constants
