+++ src/view/components/cpx/Cpx.spec.tsx
@@ -1,0 +1,29 @@
+import * as React from "react";
+import * as ReactDOM from "react-dom";
+import { IntlProvider } from "react-intl";
+import * as testRenderer from "react-test-renderer";
+import { Cpx } from "./Cpx";
+
+describe("Device component", () => {
+    it("renders correctly", () => {
+        const component = testRenderer
+            .create(
+                <IntlProvider locale="en">
+                    <Cpx />
+                </IntlProvider>
+            )
+            .toJSON();
+        expect(component).toMatchSnapshot();
+    });
+
+    it("renders without crashing", () => {
+        const div = document.createElement("div");
+        ReactDOM.render(
+            <IntlProvider locale="en">
+                <Cpx />
+            </IntlProvider>,
+            div
+        );
+        ReactDOM.unmountComponentAtNode(div);
+    });
+});
