+++ src/view/components/tab/Tab.tsx
@@ -1,6 +1,6 @@
 import { Pivot, PivotItem, PivotLinkFormat } from "office-ui-fabric-react";
 import * as React from "react";
-import { DEVICE_LIST_KEY } from "../../constants";
+import { DEVICE_LIST_KEY, CONSTANTS } from "../../constants";
 
 interface IProps {
     handleTabClick: (item?: PivotItem) => void;
@@ -13,9 +13,12 @@ export class Tab extends React.Component<IProps, any> {
                 linkFormat={PivotLinkFormat.tabs}
                 onLinkClick={handleTabClick}
             >
-                <PivotItem headerText="CPX" itemKey={DEVICE_LIST_KEY.CPX} />
                 <PivotItem
-                    headerText="Micro:bit"
+                    headerText={CONSTANTS.DEVICE_NAME.CPX}
+                    itemKey={DEVICE_LIST_KEY.CPX}
+                />
+                <PivotItem
+                    headerText={CONSTANTS.DEVICE_NAME.MICROBIT}
                     itemKey={DEVICE_LIST_KEY.MICROBIT}
                 />
             </Pivot>
