+++ src/view/container/device/Device.spec.tsx
@@ -5,8 +5,8 @@ import * as testRenderer from "react-test-renderer";
 import { DEVICE_LIST_KEY } from "../../constants";
 import { Device } from "./Device";
 
-describe("Device component", () => {
-    it("renders correctly", () => {
+describe("Device component should", () => {
+    it("render correctly", () => {
         const component = testRenderer
             .create(
                 <IntlProvider locale="en">
@@ -17,7 +17,7 @@ describe("Device component", () => {
         expect(component).toMatchSnapshot();
     });
 
-    it("renders without crashing", () => {
+    it("render without crashing", () => {
         const div = document.createElement("div");
         ReactDOM.render(
             <IntlProvider locale="en">
